+++ spring-breeze-erp-test/back/src/main/java/com/sb/erp/resv/service/ReservationServiceImpl.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Service;
 
 import com.sb.erp.res.dto.response.ResResponse;
 import com.sb.erp.res.repository.ResourceMapper;
-import com.sb.erp.resv.dto.reponse.ResvResponse;
-import com.sb.erp.resv.dto.reponse.StatsResvResponse;
 import com.sb.erp.resv.dto.request.ResvRequest;
 import com.sb.erp.resv.dto.request.ResvSearchRequest;
+import com.sb.erp.resv.dto.response.ResvResponse;
+import com.sb.erp.resv.dto.response.StatsResvResponse;
 import com.sb.erp.resv.repository.ReservationMapper;
 
 @Service
@@ -68,7 +68,7 @@ public class ReservationServiceImpl implements ReservationService {
     }
 
     @Override
-    public int delete(int revId) {
+    public int delete(long revId) {
     	return dao.delete(revId);
     }
